+++ tests/testthat/test-01-integration.R
@@ -1,7 +1,7 @@
 test_that("Discrete integration", {
   domain <- 2:5
   samplers <- 3:7
-  ips_ <- fm_int_object(
+  ips_ <- new_fm_int(
     tibble::tibble(x = 3:5, weight = rep(1, 3), .block = 1L:3L),
     name = "x"
   )
@@ -27,7 +27,7 @@ test_that("Continuous integration", {
   domain <- fm_mesh_1d(2:5)
 
   samplers <- cbind(3, 5)
-  ips_ <- fm_int_object(
+  ips_ <- new_fm_int(
     tibble::tibble(
       x = c(3:5, 3.5, 4.5),
       weight = c(1 / 6, 1 / 3, 1 / 6, 2 / 3, 2 / 3),
@@ -59,7 +59,7 @@ test_that("Continuous integration", {
   domain <- fm_mesh_1d(2:5, degree = 2)
 
   samplers <- cbind(3, 5)
-  ips_ <- fm_int_object(
+  ips_ <- new_fm_int(
     tibble::tibble(
       x = c(3:5, 3.5, 4.5),
       weight = c(1 / 6, 1 / 3, 1 / 6, 2 / 3, 2 / 3),